+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fixed deletes with routing and `async` callbacks
 - Fixed deletes with routing and `queue` callbacks
+- Fixed deprecation warnings
 
 ## 3.1.1